eighteenth migration

+++ Models/Pochild.cs
@@ -5,7 +5,7 @@
 		
 			public int Id { get; set; }
 			public string Pono { get; set; }
-			public int Itemid { get; set; }
+			public Int64 Itemid { get; set; }
 			public double Qty { get; set; }
 			public double Price { get; set; }
